pool renamed to static pool to reflect it's design, new pool interface has been added
